requirement.txt
langsmith
weaviate-client
fastapi[standard]
docling
google-genai